vorothree

Module bounds

Structs

BoundingBox

Represents an axis-aligned bounding box in 3D space.

Constants

BOX_ID_BACK

Bounding box ID for the back face, it is negative to prevent conflicts with generator IDs.

BOX_ID_BOTTOM

Bounding box ID for the bottom face, it is negative to prevent conflicts with generator IDs.

BOX_ID_FRONT

Bounding box ID for the front face, it is negative to prevent conflicts with generator IDs.

BOX_ID_LEFT

Bounding box ID for the left face, it is negative to prevent conflicts with generator IDs.

BOX_ID_RIGHT

Bounding box ID for the right face, it is negative to prevent conflicts with generator IDs.

BOX_ID_TOP

Bounding box ID for the top face, it is negative to prevent conflicts with generator IDs.