vorothree

Struct AlgorithmGrid

pub struct AlgorithmGrid {

    pub grid_res_x: usize,
    pub grid_res_y: usize,
    pub grid_res_z: usize,
    pub grid_scale_x: f64,
    pub grid_scale_y: f64,
    pub grid_scale_z: f64,
    pub grid_limit_x: f64,
    pub grid_limit_y: f64,
    pub grid_limit_z: f64,
    pub min_x: f64,
    pub min_y: f64,
    pub min_z: f64,
    pub grid_bins: Vec<Vec<usize>>,
    pub generator_bin_ids: Vec<usize>,
    pub bin_search_order: Vec<(isize, isize, isize, f64)>,

}

A spatial index based on a uniform grid.

This structure divides the 3D space into a fixed number of bins (voxels). It is generally faster than an octree for uniform distributions and allows for O(1) insertion and update operations, but may be less memory efficient for highly clustered data or very large sparse domains.

Fields
grid_res_x: usize

Number of bins along the X axis.

grid_res_y: usize

Number of bins along the Y axis.

grid_res_z: usize

Number of bins along the Z axis.

grid_scale_x: f64

Scale factor for X coordinate to grid index.

grid_scale_y: f64

Scale factor for Y coordinate to grid index.

grid_scale_z: f64

Scale factor for Z coordinate to grid index.

grid_limit_x: f64

Maximum valid index for X.

grid_limit_y: f64

Maximum valid index for Y.

grid_limit_z: f64

Maximum valid index for Z.

min_x: f64

Minimum X coordinate of the grid bounds.

min_y: f64

Minimum Y coordinate of the grid bounds.

min_z: f64

Minimum Z coordinate of the grid bounds.

grid_bins: Vec<Vec<usize>>

The grid bins, each containing a list of generator indices.

generator_bin_ids: Vec<usize>

Map from generator index to its current bin index.

bin_search_order: Vec<(isize, isize, isize, f64)>

Precomputed search order for visiting neighboring bins.

Implementations

impl AlgorithmGrid

pub fn new(nx: usize, ny: usize, nz: usize, bounds: &BoundingBox<3>) -> Self

Creates a new AlgorithmGrid with the specified dimensions and bounds.

The grid resolution (nx, ny, nz) determines the granularity of the spatial partitioning.

pub fn get_bin_index( &self, x: f64, y: f64, z: f64, bounds: &BoundingBox<3>, ) -> usize

Calculates the linear index of the bin corresponding to the given coordinates.