vorothree

Struct Tessellation

pub struct Tessellation<C: Cell, A: SpatialAlgorithm> {
    pub bounds: BoundingBox,
    pub generators: Vec<f64>,
    pub cells: Vec<C>,
    pub walls: Vec<Wall>,
    pub algorithm: A,
}

A geometry-based Voronoi tessellation that unifies the Cell, SpatialAlgorithm, and Wall traits.

Fields
bounds: BoundingBoxgenerators: Vec<f64>cells: Vec<C>walls: Vec<Wall>algorithm: A
Implementations

impl<C: Cell, A: SpatialAlgorithm> Tessellation<C, A>

pub fn new(bounds: BoundingBox, algorithm: A) -> Self

pub fn set_generators(&mut self, generators: &[f64])

Update all generators at once. Only accepts generators that are inside the bounding box and contained by the walls.

Arguments

generators - A flat array of coordinates [x, y, z, x, y, z, ...].

pub fn set_generator(&mut self, index: usize, x: f64, y: f64, z: f64)

Update the position of a single generator by index. Only sets the generator if it is inside the bounding box and contained by the walls.

pub fn random_generators(&mut self, count: usize)

Generates random points within the boundaries of the bounding box and walls and sets them as generators.

pub fn import_generators<P: AsRef<Path>>(&mut self, path: P) -> Result<()>

Imports generators from a text file. Each line should contain an id followed by N coordinate entries. For now only the first 3 coordinates are used (x, y, z) and the id is ignored.

fn prune_outside_generators(&mut self)

Removes generators that are not inside the defined walls. Note: This changes the indices of the remaining generators.

pub fn add_wall(&mut self, wall: Wall)

Adds a wall to the tessellation to clip the Voronoi cells.

pub fn clear_walls(&mut self)

Removes all walls from the tessellation.

pub fn calculate(&mut self)

Calculates all cells based on the current generators.

This method applies the SpatialAlgorithm to efficiently find the closest generators and clips the cells against the generators, the bounding box and any added walls. For the clipping it applies the algoritm as defined in the Cell implementation. It runs in parallel if the rayon feature is enabled (which is default).

pub fn map<F, T>(&self, f: F) -> Vec<T>
where F: Fn(C) -> T + Sync + Send, T: Send,

Computes cells and applies a mapping function f to each cell, returning the collected results.

This method is memory-efficient as it does not store the intermediate Cell objects. It runs in parallel if the rayon feature is enabled (which is default).

fn compute_cell( i: usize, generators: &[f64], bounds: &BoundingBox, walls: &[Wall], algorithm: &A, scratch: &mut C::Scratch, ) -> C

pub fn relax(&mut self)

Performs one step of Lloyd’s relaxation.

This moves each generator to the centroid of its calculated Voronoi cell, which tends to make the cells more uniform in size and shape. A calculation step must be invoked separately to get the new Voronoi cells.

pub fn count_generators(&self) -> usize

Returns the number of generators in the tessellation.

pub fn count_cells(&self) -> usize

Returns the number of computed cells.

pub fn get_generator(&self, index: usize) -> [f64; 3]

Retrieves the position as [f64; 3] of a generator by its index.

pub fn get_cell(&self, index: usize) -> Option<C>

Retrieves a cell by its index.

pub fn generators(&self) -> Vec<f64>

Returns a copy of all generator positions as a flat vector.

pub fn cells(&self) -> Vec<C>

Returns a copy of all computed cells.