vorothree::wall

Trait WallGeometry

pub trait WallGeometry<const D: usize>:
    Send
    + Sync
    + Debug {
    // Required methods
    fn contains(&self, point: &[f64; D]) -> bool;
    fn cut(
        &self,
        generator: &[f64; D],
        callback: &mut dyn FnMut([f64; D], [f64; D]),
    );
}

Trait defining the geometry and logic of a wall. Must be Send + Sync to support parallel execution in Tessellation.

Required Methods

fn contains(&self, point: &[f64; D]) -> bool

Checks if a point is inside the valid region defined by the wall.

fn cut( &self, generator: &[f64; D], callback: &mut dyn FnMut([f64; D], [f64; D]), )

Calculates the clipping plane for a given generator. Returns a tuple (point_on_plane, plane_normal). The normal should point OUT of the valid region (towards the region to be clipped).