vorothree::wall

Trait WallGeometry

pub trait WallGeometry:
    Send
    + Sync
    + Debug {
    // Required methods
    fn contains(&self, point: &[f64; 3]) -> bool;
    fn cut(
        &self,
        generator: &[f64; 3],
        callback: &mut dyn FnMut([f64; 3], [f64; 3]),
    );
}

Trait defining the geometry and logic of a wall. Must be Send + Sync to support parallel execution in Tessellation.

Required Methods

fn contains(&self, point: &[f64; 3]) -> bool

Checks if a point is inside the valid region defined by the wall.

fn cut( &self, generator: &[f64; 3], callback: &mut dyn FnMut([f64; 3], [f64; 3]), )

Calculates the clipping plane for a given generator. Returns a tuple (point_on_plane, plane_normal). The normal should point OUT of the valid region (towards the region to be clipped).